+++ app/Livewire/Portal/Invitation/CreateCandidateUser.php
@@ -5,14 +5,13 @@ namespace App\Livewire\Portal\Invitation;
 use App\Models\candidatePortalInvitation;
 use App\Models\Candidates;
 use App\Models\CandidateUser;
-use App\Notifications\NewCandidatePortalAccountRegisteredNotification;
+use App\Notifications\Candidates\NewCandidatePortalAccountRegisteredNotification;
 use Carbon\Carbon;
 use Filament\Actions\Action;
 use Filament\Actions\Concerns\InteractsWithActions;
 use Filament\Forms;
 use Filament\Forms\Concerns\InteractsWithForms;
 use Filament\Notifications\Notification;
-use Filament\Pages\BasePage;
 use Filament\Pages\Concerns\InteractsWithFormActions;
 use Filament\Pages\SimplePage;
 use Filament\Support\Colors\Color;
